+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+### Changed
+
+- Clear selection timeout when `SelectField` is unmounted so it doesn't leak memory.
+
 ## [13.4.0] - 2019-07-30
 
 ### Added